% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/build_image.R
\name{build_image}
\alias{build_image}
\title{Build a docker image with server that serves model$predict}
\usage{
build_image(model)
}
\description{
Available routes:
}
\details{
GET/POST          /            "OK"
GET/POST          /ping        "pong"
GET/POST          /predict     serialized output of predict as JSON
}

