% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_bin.R
\name{check_bin}
\alias{check_bin}
\title{check_bin}
\usage{
check_bin(bin_path)
}
\arguments{
\item{bin_path}{A \code{list(character)}. A list giving the binary path of
\code{vcftools}, \code{bcftools}, \code{bgzip}, \code{tabix} and \code{plink1.9}.}
}
\description{
check_bin
}
\keyword{internal}
