% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ethnicity.R
\name{estimate_ethnicity}
\alias{estimate_ethnicity}
\title{Format VCF files and compute the genomic components (and some figures) for ethnicity.}
\usage{
estimate_ethnicity(
  cohort_name,
  input_vcfs,
  input_type,
  output_directory,
  ref1kg_vcfs,
  ref1kg_population,
  ref1kg_maf = 0.05,
  splitted_by_chr = TRUE,
  quality_tag = NULL,
  quality_threshold = 0.9,
  recode = "all",
  vcf_half_call = "missing",
  n_cores = 6,
  bin_path = list(vcftools = "/usr/bin/vcftools", bcftools = "/usr/bin/bcftools", bgzip
    = "/usr/bin/bgzip", tabix = "/usr/bin/tabix", plink1.9 = "/usr/bin/plink1.9")
)
}
\arguments{
\item{cohort_name}{A \code{character}. A name to describe the studied population compared to 1,000 Genomes.}

\item{input_vcfs}{A \code{character}. A path to one or several VCFs file.}

\item{input_type}{A \code{character}. Either \code{"array"} or \code{"sequencing"}.}

\item{output_directory}{A \code{character}. The path where the data and figures is written.}

\item{ref1kg_vcfs}{A \code{character}. A path to the reference VCFs files (i.e., 1,000 Genomes sequencing data).}

\item{ref1kg_population}{A \code{character}. A file which describe samples and their ethnicity.}

\item{ref1kg_maf}{A \code{numeric}. MAF threshold for SNPs in 1,000 Genomes}

\item{splitted_by_chr}{A \code{logical}. Is the VCFs files splitted by chromosome?}

\item{quality_tag}{A \code{character}. Name of the imputation quality tag for \code{"array"},
\emph{e.g.}, \code{"INFO"} or \code{"R2"}. Default is \code{NULL}.}

\item{quality_threshold}{A \code{numeric}. The threshold to keep/discard SNPs based on their imputation quality.}

\item{recode}{A \code{character}. Which VCF should be filtered and recode, either \code{"all"} or \code{"input"}.}

\item{vcf_half_call}{A \code{character}. The mode to handle half-call.
+ 'haploid'/'h': Treat half-calls as haploid/homozygous (the PLINK 1 file format does not distinguish between the two). This maximizes similarity between the VCF and BCF2 parsers.
+ 'missing'/'m': Treat half-calls as missing (default).
+ 'reference'/'r': Treat the missing part as reference.}

\item{n_cores}{An \code{integer}. The number of CPUs to use to estimate the ethnicity.}

\item{bin_path}{A \code{list(character)}. A list giving the binary path of
\code{vcftools}, \code{bcftools}, \code{bgzip}, \code{tabix} and \code{plink1.9}.}
}
\value{
A \code{data.frame}.
}
\description{
Format VCF files and compute the genomic components (and some figures) for ethnicity.
}
