% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_sequencing.R
\name{format_sequencing}
\alias{format_sequencing}
\title{format_sequencing}
\usage{
format_sequencing(
  cohort_name,
  input_vcfs,
  output_directory,
  ref1kg_vcfs,
  ref1kg_maf,
  recode,
  vcf_half_call,
  bin_path
)
}
\arguments{
\item{cohort_name}{A \code{character}. A name to describe the studied population compared to 1,000 Genomes.}

\item{input_vcfs}{A \code{character}. A path to one or several VCFs file.}

\item{output_directory}{A \code{character}. The path where the data and figures is written.}

\item{ref1kg_vcfs}{A \code{character}. A path to the reference VCFs files (i.e., 1,000 Genomes sequencing data).}

\item{ref1kg_maf}{A \code{numeric}. MAF threshold for SNPs in 1,000 Genomes}

\item{recode}{A \code{character}. Which VCF should be filtered and recode, either \code{"all"} or \code{"input"}.}

\item{vcf_half_call}{A \code{character}. The mode to handle half-call.
+ 'haploid'/'h': Treat half-calls as haploid/homozygous (the PLINK 1 file format does not distinguish between the two). This maximizes similarity between the VCF and BCF2 parsers.
+ 'missing'/'m': Treat half-calls as missing (default).
+ 'reference'/'r': Treat the missing part as reference.}

\item{bin_path}{A \code{list(character)}. A list giving the binary path of
\code{vcftools}, \code{bcftools}, \code{bgzip}, \code{tabix} and \code{plink}.}
}
\description{
format_sequencing
}
\keyword{internal}
