% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_ethnicity_done.R
\name{is_ethnicity_done}
\alias{is_ethnicity_done}
\title{is_ethnicity_done}
\usage{
is_ethnicity_done(output_directory)
}
\arguments{
\item{output_directory}{A \code{character}. The path where the data and figures is written.}
}
\description{
is_ethnicity_done
}
\keyword{internal}
