% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kunteynir-package.R
\docType{package}
\name{kunteynir}
\alias{kunteynir}
\alias{kunteynir-package}
\title{Convert a tundraContainer into a dockerized REST server.}
\description{
Convert a tundraContainer into a dockerized REST server.
}

