% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_vcf.R
\name{merge_vcf}
\alias{merge_vcf}
\title{merge_vcf}
\usage{
merge_vcf(input_vcfs, bin_path)
}
\arguments{
\item{input_vcfs}{A \code{character}. A path to one or several VCFs file.}

\item{bin_path}{A \code{list(character)}. A list giving the binary path of
\code{vcftools}, \code{bcftools}, \code{bgzip}, \code{tabix} and \code{plink}.}
}
\description{
merge_vcf
}
\keyword{internal}
