% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_report.R
\name{pca_report}
\alias{pca_report}
\title{Compute an analysis report using principal component analysis from flashpca tool.}
\usage{
pca_report(
  data,
  design,
  id_var = "Sample_ID",
  technical_vars,
  n_comp = 10,
  fig_n_comp = n_comp,
  outliers_threshold = 1.5,
  title_level = 3
)
}
\arguments{
\item{data}{A \code{vector} or \code{data.frame}. The numeric data on which the PCA has to be performed.}

\item{design}{A \code{data.frame}. Additional variables to be used with factorial planes.}

\item{id_var}{A \code{character}. The identifier column used to merge the data.}

\item{technical_vars}{A \code{vector(character)}. Variables from design to be used with factorial planes.}

\item{n_comp}{A \code{numeric}. The number of principal components to be computed.}

\item{fig_n_comp}{A \code{numeric}. The number of principal components to be used for figures.}

\item{outliers_threshold}{A \code{numeric}. The threshold to define outliers.}

\item{title_level}{A \code{numeric}. The markdown title level, \emph{i.e.}, the number of \verb{#} preceding the section.}
}
\value{
A \code{data.frame}.
}
\description{
The function can be used in a chunk within a Rmarkdown document/script with results="asis" to render the report.
}
\examples{

if (interactive()) {
  pca_report(
    data = t(mtcars),
    design = as.data.table(mtcars, keep.rownames = "Sample_ID"),
    id_var = "Sample_ID",
    technical_vars = c("cyl", "gear", "vs"),
    n_comp = 5,
    fig_n_comp = 5,
    outliers_threshold = 3,
    title_level = 0
  )
}

}
