% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{rain-package}
\alias{rain-package}
\title{Rythmicity Analysis incooperating Non-parametric Methods}
\description{
Package Description
}
\details{
Start with function \code{'\link{rain}'}
}
\author{
Paul Thaben \email{paul.thaben@charite.de}
}

