% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pca.R
\name{compute_pca}
\alias{compute_pca}
\title{Compute the genomic components (and some figures) for ethnicity based on VCF files.}
\usage{
compute_pca(
  cohort_name,
  input_plink,
  output_directory,
  ref1kg_population,
  n_comp = 10,
  point_size = 0.75,
  label_fontsize = 5
)
}
\arguments{
\item{cohort_name}{A \code{character}. A name to describe the studied population compared to 1,000 Genomes.}

\item{input_plink}{A \code{character}. The path to plink format files (i.e., \code{.bed}, \code{.bim} and \code{.fam} files).}

\item{output_directory}{A \code{character}. The path where the data and figures is written.}

\item{ref1kg_population}{A \code{character}. A file which describe samples and their ethnicity.}

\item{n_comp}{A \code{numeric}. The number of principal components to be computed.}

\item{point_size}{A \code{numeric}. The size of the points drawn by \code{geom_point()}.}

\item{label_fontsize}{A \code{numeric}. The font size of the labels drawn by \code{geom_mark_hull()}.}
}
\value{
A \code{data.frame}.
}
\description{
Compute the genomic components (and some figures) for ethnicity based on VCF files.
}
