% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rain-package.R
\docType{package}
\name{rain-package}
\alias{rain}
\alias{rain-package}
\title{rain: Reference-based Ancestry INference}
\description{
Compute ancestry inference (principal component analysis) of a set of individuals 
    using either array or sequencing data (i.e., plink or VCF format) 
    and a reference panel (e.g., 1,000 genomes project).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mcanouil/rain}
  \item \url{https://mcanouil.github.io/rain}
  \item Report bugs at \url{https://github.com/mcanouil/rain/issues}
}

}
\author{
\strong{Maintainer}: Mickaël Canouil \email{mickael.canouil@cnrs.fr} (\href{https://orcid.org/0000-0002-3396-4549}{ORCID})

Authors:
\itemize{
  \item Mathilde Boissel \email{mathilde.boissel@cnrs.fr} (\href{https://orcid.org/0000-0001-5719-8749}{ORCID})
}

}
\keyword{internal}
