% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kunteynir-package.R
\docType{package}
\name{rain}
\alias{rain}
\alias{rain-package}
\title{Convert a tundraContainer into a dockerized REST server.}
\description{
Convert a tundraContainer into a dockerized REST server.
}

